+++ src/session/local/wait.cc
@@ -88,7 +88,7 @@
 
 	}
 
-	void Local::Session::wait(unsigned short row, unsigned short col, const char *text, int seconds) {
+	void Local::Session::wait(uint32_t row, uint32_t col, const char *text, int seconds) {
 
 		std::lock_guard<std::recursive_mutex> lock(sync);
 		chkResponse(lib3270_wait_for_string_at(hSession,row,col,convertToHost(text,-1).c_str(),seconds));
